Move modifications to central file
